+++ lib/sensirion/sensirion.hpp
@@ -9,7 +9,7 @@
 class sensirion {
   public:
     static uint8_t crc(const uint8_t byte0, const uint8_t byte1);        // calculate the CRC of two bytes according to the Sensirion CRC-8 algorithm.
-    static uint8_t generateCRCGeneric(const uint8_t* data, size_t count, uint8_t init, uint8_t polynomial); // code from Sensirion arduino library
+    static uint8_t generateCRCGeneric(const uint8_t* data, uint8_t count, uint8_t init, uint8_t polynomial); // code from Sensirion arduino library
     static bool checkCrc(const uint8_t* data, uint32_t count);        // check all CRCs in a buffer of data. Every third byte is a CRC of the previous two bytes.
     static void insertCrc(uint8_t* data, uint32_t count);             // insert CRCs in a buffer of data. Every third byte is a CRC of the previous two bytes.
     static float asFloat(const uint8_t* bytes);                       // convert a sensirion 4-byte array to a float
